requirements.txt
streamlit==1.31.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
scikit-image==0.22.0
matplotlib==3.8.2